requirements.txt
certifi==2018.1.18
diff-match-patch==20121119
Django==2.0.2
django-import-export==1.0.0
et-xmlfile==1.0.1
jdcal==1.3
mkl-fft==1.0.0
mkl-random==1.0.1
numpy==1.14.2
odfpy==1.3.6
olefile==0.45.1
openpyxl==2.5.1
pandas==0.22.0
Pillow==5.0.0
python-dateutil==2.7.2
pytz==2018.3
PyYAML==3.12
scikit-learn==0.19.1
scipy==1.0.0
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
wincertstore==0.2
xlrd==1.1.0
xlwt==1.3.0
